requirements.txt
d-geolib==2.5.0
pydantic==2.10.2
openpyxl==3.1.5